LPS-63931 Convert DDMDisplayRegistry into a Component
